change type signature of error procs
This one is a bit subtle:

The result of an `abs` call is highly dependent on the type. For a
`Vector[T]` for instance that reduces to a scalar `U`.

In case of a regular "scalar" like number, it'd remain as `T`
though. Therefore the return type is `auto` as we cannot easily write
down a proper signature.