determine type of `V` (error in `IntervalType`) from dummy call
Separate commit to explain in more words (can be squashed later).
We use a `default` based `typeof` call to determine what the type of
the errors is going to be for the given type `T`. The `default` can be
as broken as desired ( ;) ), as it's only required to determine the
return type.