force conversion of `xStart`, `xStop` using `default`
This is a pretty big hack. I'm just not sure what a better alternative
is.

The (possibly) user given type `U`, may be a different type than
`float`. And if it is an `object` type, it's not possible to construct
such a type using `U(0.0)`. And unfortunately we don't have a
constructor that always has the same name, e.g. `init[U](0.0)`.

While we could enforce such a constructor to exist as part of the
concept that will be used to match the `U` types, it doesn't exist as
of yet.

In theory this usage of `default` should be mostly safe. If the user
type doesn't handle addition with a literal, it likely won't be able
to work with these routines anyway (these are inside of comparisons
with `Inf` literals anyway, so some semblance of literal handling must
exist). And given that Nim zero initializes everything on a `default`
and zero being the identity under addition, this *should* in principle
be okay.